Moved integration tests into file that identifies them as such
