+++ src/tests/unit/tests/reports/package/reporter-factory.test.ts
@@ -2,32 +2,11 @@
 // Licensed under the MIT License.
 import { Reporter } from 'reports/package/reporter';
 import { reporterFactory } from 'reports/package/reporter-factory';
-import { scanIssues } from 'tests/unit/tests/reports/package/scans/scan-issues';
-import { scanNoIssues } from 'tests/unit/tests/reports/package/scans/scan-no-issues';
 
 describe('ReporterFactory', () => {
     it('returns a valid reporter', () => {
         const reporter = reporterFactory();
 
         expect(reporter).toBeInstanceOf(Reporter);
-    });
-
-    const options = {
-        browserVersion: 'BROWSER_VERSION',
-        browserSpec: 'BROWSER_SPEC',
-        pageTitle: 'PAGE_TITLE',
-        description: 'DESCRIPTION',
-    };
-
-    it('works end-to-end with no issues scan', () => {
-        const reporter = reporterFactory();
-        const html = reporter.fromAxeResult(scanNoIssues, options).asHTML();
-        expect(html).toMatchSnapshot();
-    });
-
-    it('works end-to-end with issues scan', () => {
-        const reporter = reporterFactory();
-        const html = reporter.fromAxeResult(scanIssues, options).asHTML();
-        expect(html).toMatchSnapshot();
     });
 });
